backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==6.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.1
PyMySQL==1.1.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
python-docx>=0.8.11
reportlab>=4.0.0
fpdf2>=2.7.5